Completed the bloc folder generation command

+++ extension.js
@@ -8,8 +8,6 @@ const vscode = require('vscode');
  */
 function activate(context) {
 
-	console.log('Congratulations, your extension "flutter-bloc-generator" is now active!');
-
 	let addToModel = vscode.commands.registerCommand('flutter-bloc-generator.addToModel', async function (uri = vscode.Uri) {
 		if (uri.fsPath.includes('state')) {
 			const openPath = vscode.Uri.file(uri.fsPath);
@@ -60,59 +58,21 @@ function activate(context) {
 	});
 
 	let blocGen = vscode.commands.registerCommand('flutter-bloc-generator.blocGen', async function (uri = vscode.Uri) {
-		const copyPath = path.join(uri.fsPath, 'copy');
-		const blocPath = path.join(copyPath, 'bloc')
-		const resPath = path.join(vscode.extensions.getExtension('DinithHerath.flutter-bloc-generator').extensionUri.fsPath, 'resources');
-		fse.copySync(resPath, copyPath);
-		for (var file of fse.readdirSync(blocPath)) {
-			fs.renameSync(path.join(blocPath, file), path.join(blocPath, file.replace('template', 'root')))
-			vscode.workspace.openTextDocument(vscode.Uri.file(path.join(blocPath, file))).then(doc => {
-				var text = doc.getText().replaceAll('Template', 'Root').replaceAll('template', 'root');
-				console.log(text);
-				fs.writeFileSync(path.join(blocPath, file), text);
-			})
+		const p = await vscode.window.showInputBox({placeHolder: 'Enter package name (as home)'});
+		const c = await vscode.window.showInputBox({placeHolder: 'Enter class name (as Home)'});
+		if (p == undefined || c == undefined) vscode.window.showInformationMessage('Cancelled adding a bloc folder');
+		else if (p == '' || c == '') vscode.window.showWarningMessage('Cannot add empty bloc folders');
+		else {
+			const copyPath = path.join(uri.fsPath, p);
+			const blocPath = path.join(copyPath, 'bloc');
+			const resPath = path.join(vscode.extensions.getExtension('DinithHerath.flutter-bloc-generator').extensionUri.fsPath, 'resources');
+			fse.copySync(resPath, copyPath);
+			var files = fse.readdirSync(blocPath);
+			for (var file of files) {
+				replaceFileWithName(blocPath, file, c, p);
+			}
+			vscode.window.showInformationMessage(`Bloc folder '${p}' created. Bloc and components folders can be found in that.`)
 		}
-		// var text = '';
-		// const p = await vscode.window.showInputBox({placeHolder: 'Enter package name (as home)'});
-		// const c = await vscode.window.showInputBox({placeHolder: 'Enter class name (as Home)'});
-		// if (p == undefined || c == undefined) {
-		// 	vscode.window.showInformationMessage('Cancelled adding bloc folders');
-		// } else if (p == '' || c == '') {
-		// 	vscode.window.showWarningMessage('Cannot add empty bloc folders');
-		// } else {
-		// 	vscode.workspace.openTextDocument(openPath).then(doc => {
-		// 		const lines = doc.getText().split('\n');
-		// 		var i = 1;
-		// 		for (let l = 0; l < lines.length; l++) {
-		// 			text += lines[l] + '\n';
-		// 			if (lines[l].trim().endsWith('{')) {
-		// 				if (i == 1) {
-		// 					i += 1;
-		// 					text += `final ${t} ${v};\n`;
-		// 				}
-		// 				else if (i == 2) {
-		// 					i += 1;
-		// 					text += `@required this.${v},\n`;
-		// 				}
-		// 				else if (i == 4) {
-		// 					i += 1;
-		// 					text += `${t} ${v},\n`;
-		// 				}
-		// 			}
-		// 			else if (lines[l].trim().endsWith('(')) {
-		// 				if (i == 3) {
-		// 					i += 1;
-		// 					text += `${v}: null,\n`;
-		// 				}
-		// 				else if (i == 5) {
-		// 					i += 1;
-		// 					text += `${v}: ${v} ?? this.${v},\n`;
-		// 				}
-		// 			}
-		// 		}
-		// 		fs.writeFileSync(uri.fsPath, text);
-		// 	});
-		// }
 	});
 
 	context.subscriptions.push(addToModel);
@@ -121,6 +81,27 @@ function activate(context) {
 
 // this method is called when your extension is deactivated
 function deactivate() { }
+
+
+/**
+ * @param {string} blocPath
+ * @param {string} [fileName]
+ * @param {string} [className]
+ * @param {string} [packageName]
+ */
+function replaceFileWithName(blocPath, fileName, className, packageName) {
+	var filePath = path.join(blocPath, fileName);
+	fs.readFile(filePath, 'utf8', function (err, data) {
+		if (err) vscode.window.showErrorMessage(err.message);
+		var text = data.replaceAll('Template', className).replaceAll('template', packageName);
+		fs.writeFile(filePath, text, function (err) {
+			if (err) vscode.window.showErrorMessage(err.message);
+			fs.rename(path.join(blocPath, fileName), path.join(blocPath, fileName.replace('template', packageName)), function (err) {
+				if (err) vscode.window.showErrorMessage(err.message);
+			});
+		});
+	});
+}
 
 module.exports = {
 	activate,
